+++ public/javascripts/reviews.js
@@ -1,0 +1,10 @@
+
+
+// const editButton = document.querySelector('#route-edit-button');
+
+// if (editButton) {
+//     const splitURL = document.URL.split('/');
+//     const cragId = splitURL[splitURL.length - 1];
+//     const postInfo = document.querySelector(`#route-post-${route.id}`)
+//     const formInfo = document.querySelector(`#route-edit-`)
+// }
